fix to Money library sum balances, add simplejson library and money

+++ source/components/responses/children.py
@@ -1,11 +1,12 @@
 from dataclasses import dataclass
+from decimal import Decimal
 from typing import Union
 
 
 @dataclass
 class DBalance:
     __slots__ = {"balance", "currency"}
-    balance: float
+    balance: Decimal
     currency: str
